sample_app/requirements.txt
requests>=2.0
dnspython>=2.0
